requirements/base.txt
Django>=4.2,<4.3
python-dotenv==1.0.1
django-tailwind==3.8.0
django-import-export==3.3.7
django-widget-tweaks==1.5.0
pillow==10.2.0
djangorestframework==3.15.1